pyproject.toml
[project]
name = "nomenclature-iamc"
authors = [
    { name = "Scenario Services team, ECE program, IIASA"},
]
description = "Package for managing codelists & attributes for IAMC-format datasets"
license = "APACHE-2.0"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pyam-iamc (>=3.1,<4.0)",
    "openpyxl (>=3.1.2,<4.0)",
    "pydantic (>=2,<3)",
    "PyYAML (>=6.0.1,<7.0)",
    "pandas (>=1.5.2,<3.0)",
    "pycountry==23.12.11",  # pin to guard against "silent" changes in country names
    "pysquirrel (>=1.1,<2.0)",
    "gitpython (>=3.1.40,<4.0)",
    "numpy (>=1.23.0,<3.0)",
    "scse-toolkit (>=0.14.0,<0.16.0)",
    "typer (>=0.20.0,<0.21.0)",
]
dynamic = ["version"]

[tool.poetry]
packages = [{include = "nomenclature"}, { include = "nomenclature/py.typed" }]
requires-poetry = ">=2.0,<3.0"
version = "0.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
mypy = "^1.7.1"
flake8 = "^6.1.0"
pytest = "^8.4.0"
coverage = "^7.3.3"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-click = "^5.1.0"
numpydoc = "^1.6.0"

[tool.ruff]
target-version="py311"

[project.scripts]
nomenclature = 'nomenclature.cli:app'

[tool.poetry-dynamic-versioning]
bump = true
enable = true
style = "pep440"
vcs = "git"

[tool.flake8]
max_line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.7,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
